requirements.txt
fastapi
uvicorn
pillow
PyMuPDF
opencv-python
pytesseract
openai
numpy
python-multipart
